% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpredEG_v0.R
\name{dpredEG_v0}
\alias{dpredEG_v0}
\title{The Exponential-Gamma Predictive Distribution}
\usage{
dpredEG_v0(ypred, y, d, dt, gm)
}
\arguments{
\item{ypred}{survival time for which predictive density is desired}

\item{y}{ordered list of observed survival times and censored survival times (y1,...,yN), yi ~ exp(theta), theta ~ Gamma(dt, gm)}

\item{d}{< N s.t. (yd+1, ..., yN) are censored}

\item{dt}{Gamma shape parameter for distribution of theta}

\item{gm}{Gamma rate parameter for distribution of theta}
}
\value{
Exponential-Gamma predictive probability
}
\description{
The Exponential-Gamma Predictive Distribution
}
\examples{
1
}
