% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpredNormIG1.R
\name{dpredNormIG1}
\alias{dpredNormIG1}
\title{The Normal-Inverse Gamma Predictive Distribution}
\usage{
dpredNormIG1(
  ypred,
  y,
  mu0 = 0,
  k0 = 1,
  sig20 = 1,
  nu0 = 1,
  S = 1e+05,
  Jeffreys = FALSE
)
}
\arguments{
\item{ypred}{vector of values for which predictive probability is desired}

\item{y}{vector of (observed) Normally-distributed values}

\item{mu0}{mean of prior observations (set default)}

\item{k0}{number of prior observations (default is 1)}

\item{sig20}{variance of prior observations (set default)}

\item{nu0}{number of prior observations (default is 1)}

\item{S}{size of random sample used for density approximation (default is 10^5)}

\item{Jeffreys}{flag for use of Jeffrey's prior}
}
\value{
random sample of size n Normal - Inverse Gamma predictive probability
}
\description{
dpredNormIG1 returns the predictive probability of future observations based on
}
\examples{
1
}
