% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpredPG - Copy.R, R/dpredPG.R
\name{dpredPG}
\alias{dpredPG}
\title{Title}
\usage{
dpredPG(x, obs, alpha = 1, beta = 1)

dpredPG(x, obs, alpha = 1, beta = 1)
}
\arguments{
\item{x}{vector of integers for which predictive probability is desired}

\item{obs}{vector of (observed) Poisson-distributed counts}

\item{alpha}{sum of counts from beta prior observations for gamma prior distribution on theta}

\item{beta}{number of prior observations for gamma prior distribution on theta}

\item{xmax}{maximum integer for which predictive probability is desired}
}
\value{
Poisson-Gamma predictive probability

Poisson-Gamma predictive probability
}
\description{
dpredPG returns the predictive probability of a future observation having value {1,...,xmax} give observations obs~Poi(theta)
and theta~Gamma(alpha,beta)

dpredPG returns the predictive probability of a future observation having value {1,...,x} given observations obs~Poi(theta)
and theta~Gamma(alpha,beta)
}
\examples{
1
1
}
