% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppredBB.R
\name{ppredBB}
\alias{ppredBB}
\title{The Beta-Binomial Predictive Distribution}
\usage{
ppredBB(x, N, s, M, alpha = 1, beta = 1, lower = TRUE)
}
\arguments{
\item{x}{vector of integers}

\item{N}{number of observed independent binary variables}

\item{s}{number of observed successes out of N}

\item{M}{number of future observations for which prediction is desired (note x <= M)}

\item{alpha}{shape factors for prior Beta distribution on theta}

\item{beta}{rate factor for prior Beta distribution on theta}

\item{lower}{logical; if TRUE (default), probabilities are P(R<=q) otherwise P(R>q)}
}
\value{
Beta-Binomial cumulative predictive probability
}
\description{
ppredBB returns the predictive probability of <= x future successes out of M trials, given s observed successes out of a sample of
}
\examples{
1
}
