% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppredEG.R
\name{ppredEG}
\alias{ppredEG}
\title{The Exponential-Gamma Predictive Distribution}
\usage{
ppredEG(x, XN, d, dt, gm)
}
\arguments{
\item{x}{survival time}

\item{XN}{(x1,...,xN), xi ~ exp(theta), theta ~ Gamma(dt, gm)}

\item{d}{< n s.t. (x1,...,xd) are fully observed and (xd+1, ..., xN) are censored}

\item{dt}{Gamma shape parameter for distribution of theta}

\item{gm}{Gamma rate parameter for distribution of theta}
}
\value{
Exponential-Gamma cumulative predictive probability
}
\description{
The Exponential-Gamma Predictive Distribution
}
\examples{
1
}
