% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppredPG.R
\name{ppredPG}
\alias{ppredPG}
\title{The Poisson-Gamma Predictive Distribution}
\usage{
ppredPG(ypred, y, alpha = 1, beta = 1, lower = TRUE)
}
\arguments{
\item{ypred}{non-negative integer count for which predictive probability is desired}

\item{y}{vector of (observed) Poisson-distributed counts}

\item{alpha}{sum of counts from beta prior observations for gamma prior distribution on theta}

\item{beta}{number of prior observations for gamma prior distribution on theta}

\item{lower}{logical; if TRUE (default), probabilities are P(R<=q) otherwise P(R>q)}
}
\value{
Poisson-Gamma cumulative predictive probability
}
\description{
ppredPG returns the cumulative predictive probability of future observation having value {1,...,xmax} given observations obs~Poi(theta)
and theta~Gamma(alpha,beta)
}
\examples{
1
}
