% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredBB.R
\name{rpredBB}
\alias{rpredBB}
\title{The Beta-Binomial Predictive Distribution}
\usage{
rpredBB(S, N, t, M, alpha = 1, beta = 1)
}
\arguments{
\item{S}{desired random sample size}

\item{N}{number of observed independent binary variables}

\item{t}{number of observed successes out of N observations}

\item{M}{number of future observations for which prediction is desired}

\item{alpha}{shape factor for prior Beta distribution on theta}

\item{beta}{rate factor for prior Beta distribution on theta}
}
\value{
random sample of size n from the Beta-Binomial predictive probability distribution
}
\description{
rpredBB returns a random sample of size n from the Beta-Binomial predictive probability distribution computed
from s observed successes out of a sample of size N, M future observations, and user input shape parameters for Beta prior on Pr(success)
}
\examples{
1
}
