% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredEG.R
\name{rpredEG}
\alias{rpredEG}
\title{Title}
\usage{
rpredEG(n, XN, d, dt, gm)
}
\arguments{
\item{n}{desired random sample size}

\item{XN}{(x1,...,xN), xi ~ exp(theta), theta ~ Gamma(dt, gm)}

\item{d}{< n s.t. (x1,...,xd) are fully observed and (xd+1, ..., xN) are censored}

\item{dt}{Gamma shape parameter for distribution of theta}

\item{gm}{Gamma rate parameter for distribution of theta}
}
\value{
random sample of size n from the Exponential-Gamma predictive probability distribution
}
\description{
Title
}
\examples{
1
}
