% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredEG_v0.R
\name{rpredEG_v0}
\alias{rpredEG_v0}
\title{The Exponential-Gamma Predictive Distribution}
\usage{
rpredEG_v0(S, y, d, dt, gm)
}
\arguments{
\item{S}{desired random sample size}

\item{y}{list of observed and censored survival times (y1,...,yN), yi ~ exp(theta), theta ~ Gamma(dt, gm)}

\item{d}{< N s.t. (yd+1, ..., yN) are censored}

\item{dt}{Gamma shape parameter for distribution of theta}

\item{gm}{Gamma rate parameter for distribution of theta}
}
\value{
random sample of size S from the Exponential-Gamma predictive probability distribution
}
\description{
The Exponential-Gamma Predictive Distribution
}
\examples{
1
}
