% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredNormIG.R
\name{rpredNormIG}
\alias{rpredNormIG}
\title{Title}
\usage{
rpredNormIG(n, obs, mu0 = 1.9, k0 = 1, s20 = 0.01, nu0 = 1, Jeffreys = FALSE)
}
\arguments{
\item{n}{desired random sample size}

\item{obs}{vector of (observed) Poisson-distributed counts}

\item{mu0}{mean of prior observations (set default)}

\item{k0}{number of prior observations (default is 1)}

\item{s20}{variance of prior observations (set default)}

\item{nu0}{number of prior observations (default is 1)}

\item{Jeffreys}{flag for use of Jeffrey's prior}
}
\value{
random sample of size n Normal - Inverse Gamma predictive probability
}
\description{
rpredPG returns a random sample of size n from the Poisson-Gamma predictive probability distribution given observations obs~Poi(theta)
and theta~Gamma(alpha,beta)
}
\examples{
1
}
