% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredNormIG2.R
\name{rpredNormIG2}
\alias{rpredNormIG2}
\title{The Normal-Inverse Gamma Predictive Distribution for Two Samples}
\usage{
rpredNormIG2(
  n = 1,
  y1,
  y2,
  mu0 = 0,
  g20 = 1,
  d0 = 0,
  t20 = 1,
  s20 = 1,
  nu0 = 1
)
}
\arguments{
\item{n}{desired random sample size (default is 1)}

\item{y1}{vector of (observed) Normally-distributed values (sample 1)}

\item{y2}{vector of (observed) Normally-distributed values (sample 2)}

\item{mu0}{mean of pooled average from prior knowledge (default is 0)}

\item{g20}{variance of pooled average from prior knowledge (default is 1)}

\item{d0}{mean of delta from prior knowledge, where delta is half the population difference in means (default is 0)}

\item{t20}{variance of delta from prior knowledge (default is 1)}

\item{s20}{variance of sample from prior knowledge (default is 1)}

\item{nu0}{number of prior observations (default is 1)}
}
\value{
random sample of size n Normal - Inverse Gamma predictive probability for each population sample
}
\description{
rpredNormIG2 returns a random sample of size n from the Normal-Inverse Gamma predictive probability distribution for 2 samples
}
\examples{
1
}
