% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredNormIGk.R
\name{rpredNormIGk}
\alias{rpredNormIGk}
\title{The Normal-Inverse Gamma Predictive Distribution for k Samples}
\usage{
rpredNormIGk(S = 1, Y, nu0 = 1, s20 = 1, eta0 = 1, t20 = 1, mu0 = 0, g20 = 1)
}
\arguments{
\item{S}{desired random sample size}

\item{Y}{2-column matrix of multiple samples of observed Normally-distributed values.  Column 1:  integer indices identifying the samples.  Column 2:  data values for the sample indicated by the corresponding index in column 1.}

\item{nu0}{number of prior observations}

\item{s20}{within-sample variance from prior knowledge}

\item{eta0}{number of prior observations}

\item{t20}{between-sample variance from prior knowledge}

\item{mu0}{mean of pooled average from prior knowledge}

\item{g20}{variance of pooled average from prior knowledge}
}
\value{
random sample of size S Normal - Inverse Gamma predictive probability for each population sample
}
\description{
rpredNormIGk returns a random sample of size N from the Normal-Inverse Gamma predictive probability distribution for each of the k samples, along with N posterior draws of the within-group and between-groups unknown parameters.
}
\examples{
1
}
