% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredNormReg.R
\name{rpredNormReg}
\alias{rpredNormReg}
\title{Normal Regression Predictive Inference}
\usage{
rpredNormReg(N = 1, X, y, nu0 = 1, s20 = 1)
}
\arguments{
\item{N}{desired random sample size}

\item{X}{Explanatory variables matrix}

\item{y}{Vector of prior observations}

\item{nu0}{number of prior observations}

\item{s20}{ols estimate of variance \link{perform regression and compute this instead of user input?}}
}
\value{
random sample of size ? from ? predictive probability distribution
}
\description{
rpredNormReg returns a random sample of size ? from ? predictive probability distribution
}
\examples{
1
}
