% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredPG.R
\name{rpredPG}
\alias{rpredPG}
\title{Title}
\usage{
rpredPG(n, obs, xmax, alpha = 1, beta = 1)
}
\arguments{
\item{n}{desired random sample size}

\item{obs}{vector of (observed) Poisson-distributed counts}

\item{xmax}{maximum integer for which predictive probability is desired}

\item{alpha}{sum of counts from beta prior observations for gamma prior distribution on theta}

\item{beta}{number of prior observations for gamma prior distribution on theta}
}
\value{
Poisson-Gamma cumulative predictive probability
}
\description{
rpredPG returns a random sample of size n from the Poisson-Gamma predictive probability distribution given observations obs~Poi(theta)
and theta~Gamma(alpha,beta)
}
\examples{
1
}
