% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpredBB.R
\name{dpredBB}
\alias{dpredBB}
\title{Title}
\usage{
dpredBB(x, N, s, M, alpha = 1, beta = 1)
}
\arguments{
\item{x}{vector of integers}

\item{N}{number of observed independent binary variables}

\item{s}{number of observed successes out of N}

\item{M}{number of future observations for which prediction is desired (note x <= M)}

\item{alpha}{shape factors for prior Beta distribution on theta}

\item{beta}{rate factor for prior Beta distribution on theta}
}
\value{
predictive probability
}
\description{
dpredBB returns the predictive probability of x future successes out of M trials, given s observed successes out of a sample of size
N, and user input shape parameters for Beta prior on Pr(success)
}
\examples{
1
}
