% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppredEG.R
\name{ppredEG}
\alias{ppredEG}
\title{The Exponential-Gamma Predictive Distribution}
\usage{
ppredEG(ypred, y, c, dt, gm)
}
\arguments{
\item{ypred}{survival time(s) for which predictive density is desired}

\item{y}{ordered list of observed survival times and censored survival times (y1,...,yN), yi ~ exp(theta), theta ~ Gamma(dt, gm)}

\item{c}{censoring indicator (vector of 0s and 1s for censored and fully observed, respectively)}

\item{dt}{Gamma shape parameter for distribution of theta}

\item{gm}{Gamma rate parameter for distribution of theta}
}
\value{
Exponential-Gamma cumulative predictive probability
}
\description{
The Exponential-Gamma Predictive Distribution
}
\examples{
1
}
