% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppredPG - Copy.R, R/ppredPG.R
\name{ppredPG}
\alias{ppredPG}
\title{Title}
\usage{
ppredPG(x, obs, alpha = 1, beta = 1, lower = TRUE)

ppredPG(x, obs, alpha = 1, beta = 1, lower = TRUE)
}
\arguments{
\item{x}{non-negative integer count for which predictive probability is desired}

\item{obs}{vector of (observed) Poisson-distributed counts}

\item{alpha}{sum of counts from beta prior observations for gamma prior distribution on theta}

\item{beta}{number of prior observations for gamma prior distribution on theta}

\item{lower}{logical; if TRUE (default), probabilities are P(R<=q) otherwise P(R>q)}

\item{xmax}{maximum integer for which predictive probability is desired}
}
\value{
Poisson-Gamma cumulative predictive probability

Poisson-Gamma cumulative predictive probability
}
\description{
ppredBB returns the predictive probability of <= x future successes out of M trials, given s observed successes out of a sample of

ppredPG returns the cumulative predictive probability of future observation having value {1,...,xmax} given observations obs~Poi(theta)
and theta~Gamma(alpha,beta)
}
\examples{
1
1
}
