% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredBB.R
\name{rpredBB}
\alias{rpredBB}
\title{Title}
\usage{
rpredBB(n, N, s, M, alpha = 1, beta = 1)
}
\arguments{
\item{n}{desired random sample size}

\item{N}{number of observed independent binary variables}

\item{s}{number of observed successes out of N}

\item{M}{number of future observations for which prediction is desired (note x <= M)}

\item{alpha}{shape factors for prior Beta distribution on theta}

\item{beta}{rate factor for prior Beta distribution on theta}
}
\value{
random sample of size n from the Beta-Binomial predictive probability distribution
}
\description{
rpredBB returns a random sample of size n from the Beta-Binomial predictive probability distribution computed from s observed successes out of a sample of size N, M future observations, and user input shape parameters for Beta prior on Pr(success)
}
\examples{
1
}
