% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredNormIG.R
\name{rpredNormIG}
\alias{rpredNormIG}
\title{The Normal-Inverse Gamma Predictive Distribution}
\usage{
rpredNormIG(n, obs, mu0 = 0, k0 = 1, sig20 = 1, nu0 = 1, Jeffreys = FALSE)
}
\arguments{
\item{n}{desired random sample size}

\item{obs}{vector of (observed) Normally-distributed values}

\item{mu0}{mean of prior observations (set default)}

\item{k0}{number of prior observations (default is 1)}

\item{sig20}{variance of prior observations (set default)}

\item{nu0}{number of prior observations (default is 1)}

\item{Jeffreys}{flag for use of Jeffrey's prior}
}
\value{
random sample of size n Normal - Inverse Gamma predictive probability
}
\description{
rpredNormIG returns a random sample of size n from the Normal-Inverse Gamma predictive probability distribution
}
\examples{
1
}
