% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpredNormReg.R
\name{rpredNormReg}
\alias{rpredNormReg}
\title{Normal Regression Predictive Distribution}
\usage{
rpredNormReg(
  S = 1,
  Xpred,
  X,
  y,
  beta0,
  Sigma0,
  nu0 = 1,
  s20 = 1,
  gprior = TRUE
)
}
\arguments{
\item{S}{desired random sample size}

\item{Xpred}{predictor vector(s) for which predictions are desired}

\item{X}{Explanatory variables matrix}

\item{y}{Vector of prior observations}

\item{beta0}{vector of prior regression coefficients}

\item{Sigma0}{prior variance-covariance matrix}

\item{nu0}{number of prior observations}

\item{s20}{prior variance}

\item{gprior}{flag to use Zellner's g-prior.  Default to TRUE}
}
\value{
random sample of size ? from ? predictive probability distribution
}
\description{
rpredNormReg returns a random sample of size ? from ? predictive probability distribution
}
\examples{
1
}
